// src/app/models/reportInterface.ts
export interface Report {
  aliado: string;
  apellido: string;
  banco: string;
  email:string;
  id_pago: number;
  id_poliza: null;
  metodo_pago: string;
  fecha_emision: Date;
  monto_pago: string;
  monto:string;
  nombre: string;
  plan:string;
  referencia: string;
  status:string;
  sypago:any;
  telefono: string;
  estado:string,
  empresa:string,
  estado_poliza:string,
  aseguradora:string,
  mostrarDetalles: boolean
}